requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.7.4
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
clr-loader==0.2.6
Django==4.2.13
django-ckeditor==6.7.1
django-constance==4.1.2
django-cors-headers==4.4.0
django-filter==24.2
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-modeltranslation==0.19.6
django-nested-inline==0.4.6
django-unfold==0.39.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
drf-spectacular==0.27.2
et-xmlfile==1.1.0
everapi==0.1.1
freecurrencyapi==0.1.0
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
numpy==2.1.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pythonnet==3.0.3
pytz==2024.2
PyYAML==6.0.1
redis==5.1.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
